requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langgraph==0.0.62
langchain-core>=0.2,<0.3
langgraph==0.0.62
google-api-python-client==2.100.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-multipart==0.0.6
python-dateutil==2.8.2
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
google-genai==0.8.0
pytz==2023.3
